requirements.txt
# Core Dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Sources
polygon-api-client>=1.13.0
finnhub-python>=2.4.0
yfinance>=0.2.0

# Data Orchestration
prefect>=2.14.0
prefect-aws>=0.3.0

# Storage
boto3>=1.28.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
neo4j>=5.15.0

# Time-Series Database
# TimescaleDB is a PostgreSQL extension, so psycopg2-binary (already listed above) is sufficient
# No Python package needed - TimescaleDB runs as a PostgreSQL extension

# Feature Store
feast>=0.36.0

# Technical Analysis
TA-Lib>=0.4.28
# Note: pandas-ta is not used in the codebase (we use talib/TA-Lib instead)

# Deep Learning
torch>=2.0.0
pytorch-forecasting>=1.0.0
torch-geometric>=2.4.0
pytorch-lightning>=2.1.0

# Machine Learning
lightgbm>=4.1.0
xgboost>=2.0.0
statsmodels>=0.14.0
arch>=6.2.0

# LLM Agents
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0
sentence-transformers>=2.2.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Monitoring
mlflow>=2.9.0
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
joblib>=1.3.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0

